+++ internal/application/order/usecase/order.usecase_test.go
@@ -14,7 +14,7 @@ import (
 	"github.com/goodone-dev/go-boilerplate/internal/domain/product"
 	productmock "github.com/goodone-dev/go-boilerplate/internal/domain/product/mocks"
 	"github.com/goodone-dev/go-boilerplate/internal/infrastructure/logger"
-	busmock "github.com/goodone-dev/go-boilerplate/internal/infrastructure/message/bus/mocks"
+	busmock "github.com/goodone-dev/go-boilerplate/internal/infrastructure/messaging/bus/mocks"
 	"github.com/google/uuid"
 	"github.com/stretchr/testify/assert"
 	"github.com/stretchr/testify/mock"
