+++ internal/infrastructure/messaging/rabbitmq/IMPLEMENTATION_SUMMARY.md
