+++ internal/infrastructure/messaging/rabbitmq/client.go
