+++ internal/infrastructure/messaging/rabbitmq/errors.go
