+++ internal/infrastructure/messaging/rabbitmq/rpc/client.go
@@ -7,7 +7,7 @@ import (
 	"sync"
 	"time"
 
-	"github.com/goodone-dev/go-boilerplate/internal/infrastructure/message/rabbitmq"
+	"github.com/goodone-dev/go-boilerplate/internal/infrastructure/messaging/rabbitmq"
 	"github.com/google/uuid"
 	amqp "github.com/rabbitmq/amqp091-go"
 )
