+++ internal/presentation/messaging/bus/bus.go
@@ -3,7 +3,7 @@ package bus
 import (
 	mailhandler "github.com/goodone-dev/go-boilerplate/internal/application/mail/delivery/messaging"
 	"github.com/goodone-dev/go-boilerplate/internal/domain/mail"
-	"github.com/goodone-dev/go-boilerplate/internal/infrastructure/message/bus"
+	"github.com/goodone-dev/go-boilerplate/internal/infrastructure/messaging/bus"
 	"github.com/goodone-dev/go-boilerplate/internal/presentation/messaging/middleware"
 )
